+++ contracts/index.ts
@@ -1,3 +1,4 @@
+export * from "./analyticsEvents";
 export * from "./gameCatalog";
 export * from "./roomMatch";
 export * from "./userProfile";